OAuth/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sumsam Ali", email = "sumsamali.cs@gmail.com"},
]
dependencies = [
    "cachetools>=5.5.2",
    "jsonpath-ng>=1.6.1",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.1.20",
    "langgraph>=0.1.19",
    "pydantic-settings>=2.4.0",
    "pydantic>=2.8.2",
    "pymongo>=4.13.0",
    "python-jose>=3.4.0",
    "pytz>=2024.1",
    "rapidfuzz>=3.9.5",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.35",
    "starlette>=0.46.0",
    "xxhash>=3.4.1",
    "uvicorn>=0.35.0",
    "FastAPI>=0.116.1",
    "sqlmodel>=0.0.24",
    "aiosqlite>=0.21.0",
    "passlib>=1.7.4",
    "jwt>=1.4.0",
    "cryptography>=45.0.7",
    "pyotp>=2.9.0",
    "qrcode>=8.2",
    "opentelemetry-sdk>=1.36.0",
    "azure-monitor-opentelemetry>=1.8.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
